feat: Elasticsearch 增加日期转换

+++ common/src/main/java/com/example/entity/ResultEntity.java
@@ -24,7 +24,7 @@ public class ResultEntity<T> {
         OK = r;
     }
 
-    private long code;
+    private int code;
 
     private String message;
